From Sam Stephenson <sam@37signals.com>: We're using your vPim library at 37signals for the Backpack Calendar (http://backpackit.com/calendar ) and ran into an issue with the "US Holidays" calendar available at http://ical.mac.com/ical/US32Holidays.ics . Specifically, calling Vpim::Rrule#each for events such as Labor Day:
  DTSTART;VALUE=DATE:20020902
  DTEND;VALUE=DATE:20020903
  RRULE:FREQ=YEARLY;INTERVAL=1;BYMONTH=9;BYDAY=1MO

would never yield any recurrences.

The problem is that byday_in_yearly calls byday_in_monthly with a nil
month, which Ruby treats as January.  So when BYMONTH and BYDAY are
specified in a rule with FREQ=YEARLY, recurrences for any month other
than January won't be generated.  I fixed this by replacing the call
to byday_in_yearly with a flattened mapping of BYMONTH months to
byday_in_monthly calls.

Attached is a tested patch against the "stable" CVS module from
Rubyforge.  Let me know if it's unclear or if it doesn't work for you.